// src/global/enseignant.d.ts
export interface Enseignant{
    noEnseignant: number;
    adresse: string ;
    codePostal: number;
    emailPerso: string ;
    emailUbo: string;
    mobile: number;
    nom: string;
    pays: string;
    prenom: string;
    sexe: string;
    telephone: number;
    type: string;
    ville: string;
}